+++ inst/include/epiworld/models/sirmixing.hpp
@@ -16,7 +16,9 @@
  * @ingroup mixing_models
  */
 template<typename TSeq = EPI_DEFAULT_TSEQ>
-class ModelSIRMixing : public Model<TSeq>
+class ModelSIRMixing :
+    public Model<TSeq>,
+    public ContactMatrix
 {
 private:
 
@@ -38,7 +40,6 @@ private:
         );
 
     std::vector< double > adjusted_contact_rate;
-    std::vector< double > contact_matrix;
 
     size_t index(size_t i, size_t j, size_t n) {
         return j * n + i;
@@ -93,12 +94,6 @@ public:
         return n_infected_per_group[group];
     }
 
-    void set_contact_matrix(std::vector< double > cmat)
-    {
-        contact_matrix = cmat;
-        return;
-    };
-
 };
 
 template<typename TSeq>
@@ -151,9 +146,7 @@ inline size_t ModelSIRMixing<TSeq>::sample_agents(
         // How many from this entity?
         int nsamples = this->rbinom(
             group_size,
-            adjusted_contact_rate[g] * contact_matrix[
-                index(agent_group_id, g, ngroups)
-            ]
+            adjusted_contact_rate[g] * this->get_contact_rate(agent_group_id, g, false)
         );
 
         if (nsamples == 0)
@@ -205,27 +198,7 @@ inline void ModelSIRMixing<TSeq>::reset()
 
     // Checking contact matrix dimensions
     size_t nentities = this->entities.size();
-    if (this->contact_matrix.size() !=  nentities*nentities)
-        throw std::length_error(
-            std::string("The contact matrix must be a square matrix of size ") +
-            std::string("nentities x nentities. ") +
-            std::to_string(this->contact_matrix.size()) +
-            std::string(" != ") + std::to_string(nentities*nentities) +
-            std::string(".")
-            );
-
-    for (size_t i = 0u; i < this->entities.size(); ++i)
-    {
-        for (size_t j = 0u; j < this->entities.size(); ++j)
-        {
-            if (this->contact_matrix[index(i, j, nentities)] < 0.0)
-                throw std::range_error(
-                    std::string("The contact matrix must be non-negative. ") +
-                    std::to_string(this->contact_matrix[index(i, j, nentities)]) +
-                    std::string(" < 0.")
-                    );
-        }
-    }
+    this->validate_contact_matrix(nentities);
 
     // Do it the first time only
     sampled_agents.resize(this->size());
@@ -298,7 +271,7 @@ inline ModelSIRMixing<TSeq>::ModelSIRMixing(
 {
 
     // Setting up the contact matrix
-    this->contact_matrix = contact_matrix;
+    this->set_contact_matrix(contact_matrix);
 
     UpdateFun<TSeq> update_susceptible = [](
         Agent<TSeq> * p, Model<TSeq> * m