+++ src/cpp11.cpp
@@ -870,6 +870,20 @@ extern "C" SEXP _epiworldR_draw_mermaid_cpp(SEXP model, SEXP fn_output, SEXP sel
     return R_NilValue;
   END_CPP11
 }
+// model.cpp
+std::vector<double> get_contact_matrix_cpp(SEXP model);
+extern "C" SEXP _epiworldR_get_contact_matrix_cpp(SEXP model) {
+  BEGIN_CPP11
+    return cpp11::as_sexp(get_contact_matrix_cpp(cpp11::as_cpp<cpp11::decay_t<SEXP>>(model)));
+  END_CPP11
+}
+// model.cpp
+SEXP set_contact_matrix_cpp(SEXP model, std::vector<double> contact_matrix);
+extern "C" SEXP _epiworldR_set_contact_matrix_cpp(SEXP model, SEXP contact_matrix) {
+  BEGIN_CPP11
+    return cpp11::as_sexp(set_contact_matrix_cpp(cpp11::as_cpp<cpp11::decay_t<SEXP>>(model), cpp11::as_cpp<cpp11::decay_t<std::vector<double>>>(contact_matrix)));
+  END_CPP11
+}
 // model_builder.cpp
 SEXP Model_cpp();
 extern "C" SEXP _epiworldR_Model_cpp() {
@@ -896,6 +910,13 @@ SEXP update_fun_rate_cpp(cpp11::strings param_names, cpp11::integers target_stat
 extern "C" SEXP _epiworldR_update_fun_rate_cpp(SEXP param_names, SEXP target_states) {
   BEGIN_CPP11
     return cpp11::as_sexp(update_fun_rate_cpp(cpp11::as_cpp<cpp11::decay_t<cpp11::strings>>(param_names), cpp11::as_cpp<cpp11::decay_t<cpp11::integers>>(target_states)));
+  END_CPP11
+}
+// model_builder.cpp
+SEXP set_state_function_cpp(SEXP model, std::string state_label, SEXP update_fun);
+extern "C" SEXP _epiworldR_set_state_function_cpp(SEXP model, SEXP state_label, SEXP update_fun) {
+  BEGIN_CPP11
+    return cpp11::as_sexp(set_state_function_cpp(cpp11::as_cpp<cpp11::decay_t<SEXP>>(model), cpp11::as_cpp<cpp11::decay_t<std::string>>(state_label), cpp11::as_cpp<cpp11::decay_t<SEXP>>(update_fun)));
   END_CPP11
 }
 // network.cpp
@@ -1333,6 +1354,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_epiworldR_get_all_sample_kernel_scores_cpp",     (DL_FUNC) &_epiworldR_get_all_sample_kernel_scores_cpp,      1},
     {"_epiworldR_get_all_sample_params_cpp",            (DL_FUNC) &_epiworldR_get_all_sample_params_cpp,             1},
     {"_epiworldR_get_all_sample_stats_cpp",             (DL_FUNC) &_epiworldR_get_all_sample_stats_cpp,              1},
+    {"_epiworldR_get_contact_matrix_cpp",               (DL_FUNC) &_epiworldR_get_contact_matrix_cpp,                1},
     {"_epiworldR_get_current_accepted_params_cpp",      (DL_FUNC) &_epiworldR_get_current_accepted_params_cpp,       1},
     {"_epiworldR_get_current_accepted_stats_cpp",       (DL_FUNC) &_epiworldR_get_current_accepted_stats_cpp,        1},
     {"_epiworldR_get_current_proposed_params_cpp",      (DL_FUNC) &_epiworldR_get_current_proposed_params_cpp,       1},
@@ -1399,6 +1421,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_epiworldR_run_lfmcmc_cpp",                       (DL_FUNC) &_epiworldR_run_lfmcmc_cpp,                        5},
     {"_epiworldR_run_multiple_cpp",                     (DL_FUNC) &_epiworldR_run_multiple_cpp,                      8},
     {"_epiworldR_set_agents_data_cpp",                  (DL_FUNC) &_epiworldR_set_agents_data_cpp,                   3},
+    {"_epiworldR_set_contact_matrix_cpp",               (DL_FUNC) &_epiworldR_set_contact_matrix_cpp,                2},
     {"_epiworldR_set_death_reduction_cpp",              (DL_FUNC) &_epiworldR_set_death_reduction_cpp,               2},
     {"_epiworldR_set_death_reduction_fun_cpp",          (DL_FUNC) &_epiworldR_set_death_reduction_fun_cpp,           3},
     {"_epiworldR_set_death_reduction_ptr_cpp",          (DL_FUNC) &_epiworldR_set_death_reduction_ptr_cpp,           3},
@@ -1429,6 +1452,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_epiworldR_set_recovery_enhancer_fun_cpp",        (DL_FUNC) &_epiworldR_set_recovery_enhancer_fun_cpp,         3},
     {"_epiworldR_set_recovery_enhancer_ptr_cpp",        (DL_FUNC) &_epiworldR_set_recovery_enhancer_ptr_cpp,         3},
     {"_epiworldR_set_simulation_fun_cpp",               (DL_FUNC) &_epiworldR_set_simulation_fun_cpp,                2},
+    {"_epiworldR_set_state_function_cpp",               (DL_FUNC) &_epiworldR_set_state_function_cpp,                3},
     {"_epiworldR_set_stats_names_cpp",                  (DL_FUNC) &_epiworldR_set_stats_names_cpp,                   2},
     {"_epiworldR_set_summary_fun_cpp",                  (DL_FUNC) &_epiworldR_set_summary_fun_cpp,                   2},
     {"_epiworldR_set_susceptibility_reduction_cpp",     (DL_FUNC) &_epiworldR_set_susceptibility_reduction_cpp,      2},